+++ src/frequenz/channels/util/__init__.py
@@ -17,6 +17,9 @@ A module with several utilities to work with channels:
   multiple receivers into a single named stream, allowing to identify the
   origin of each message.
 
+* [PeriodicTimer][frequenz.channels.util.PeriodicTimer]:
+  A [receiver][frequenz.channels.Receiver] that ticks at regular intervals.
+
 * [Select][frequenz.channels.util.Select]: A helper to select the next
   available message for each [receiver][frequenz.channels.Receiver] in a group
   of receivers.
@@ -29,7 +32,7 @@ A module with several utilities to work with channels:
 from ._file_watcher import FileWatcher
 from ._merge import Merge
 from ._merge_named import MergeNamed
-from ._periodic_timer import PeriodicTimer
+from ._periodic_timer import MissedTickBehavior, PeriodicTimer
 from ._select import Select
 from ._timer import Timer
 
@@ -37,6 +40,7 @@ __all__ = [
     "FileWatcher",
     "Merge",
     "MergeNamed",
+    "MissedTickBehavior",
     "PeriodicTimer",
     "Select",
     "Timer",